hw/cxl/cxl-component-utils: Drop problematic QEMU_BUILD_BUG_ON()
It's not clear why the code is checking for a 0 offset.
Whilst one is never seen, some compilers are tripping up
on this check. Given this can be worked around by changing the
scope, it appears to be a compiler limitation rather than an actual
bug. Simplify matters by dropping the check.

Reported-by: Jeongtae Park <jtp.park@samsung.com>
Signed-off-by: Jonathan Cameron <jonathan.cameron@huawei.com>